requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
Flask-Migrate==4.0.0
Flask-Cors==3.0.10
psycopg2-binary==2.9.3
python-dotenv=1.0.1